IPD-40 #comment refactored
