+++ authorizers/iconik/interfaces/context.ts
@@ -1,3 +1,5 @@
+import { IconikParams } from '@workflowwin/iconik-api';
+
 export interface IconikContext {
   appId: string;
   systemDomainId: string;
@@ -11,4 +13,13 @@ export interface IconikContext {
   };
   authToken: string;
   iconikUrl: string;
+}
+
+export interface IconikEnhancedAuthContext extends IconikParams {
+  principalId?: string;
+  integrationLatency?: string;
+  callerId?: string;
+  callerEmail?: string;
+  ownerId?: string;
+  ownerEmail?: string;
 }
